+++ environment.yml
@@ -1,37 +1,32 @@
 name: hdw-thesis
 channels:
 - defaults
+- conda-forge
 dependencies:
-- ca-certificates=2017.08.26=ha1e5d58_0
-- certifi=2017.7.27.1=py36hd973bb6_0
-- intel-openmp=2018.0.0=h68bdfb3_7
-- libcxx=4.0.1=h579ed51_0
-- libcxxabi=4.0.1=hebd6815_0
-- libedit=3.1=hb4e282d_0
-- libffi=3.2.1=h475c297_4
-- libgfortran=3.0.1=h93005f0_2
-- mkl=2018.0.0=h5ef208c_6
-- ncurses=6.0=ha932d30_1
-- numpy=1.13.3=py36h2cdce51_0
-- openssl=1.0.2m=h86d3e6a_0
-- pandas=0.21.0=py36hfed917e_1
-- pip=9.0.1=py36h1555ced_4
-- python=3.6.3=hc655967_3
-- python-dateutil=2.6.1=py36h86d2abb_1
-- pytz=2017.2=py36h2e7dfbc_1
-- readline=7.0=h5003dca_4
-- scipy=1.0.0=py36h1de22e9_0
-- setuptools=36.5.0=py36h2134326_0
-- six=1.11.0=py36h0e22d5e_1
-- sqlite=3.20.1=h7e4c145_2
-- tk=8.6.7=h35a86e2_3
-- wheel=0.29.0=py36h3597b6d_1
-- xz=5.2.3=h0278029_2
-- zlib=1.2.11=hf3cbc9b_2
-- pip:
-  - pbr==3.1.1
-  - stevedore==1.25.0
-  - virtualenv-clone==0.2.6
-  - virtualenvwrapper==4.7.2
+ - python=3
+ - numpy
+ - h5py
+ - scipy
+ - toolz
+ - bokeh
+ - dask
+ - distributed
+ - notebook
+ - matplotlib
+ - seaborn
+ - Pillow
+ - pandas
+ - pandas-datareader
+ - pytables
+ - scikit-learn
+ - snakeviz
+ - ujson
+ - graphviz
+ - pip
+ - s3fs
+ - fastparquet
+ - pip:
+         - graphviz 
+         - cachey
 prefix: /Users/henrywilde/anaconda3/envs/hdw-thesis
