+++ thesis.bib
@@ -1,0 +1,15 @@
+% Encoding: UTF-8
+
+@Article{supervised-ml-comparison,
+  author  = {Amancio DR, Comin CH, Casanova D, Travieso G, Bruno OM, Rodrigues FA, et al.},
+  title   = {A Systematic Comparison of Supervised Classifiers},
+  journal = {PLOS ONE},
+  year    = {2014},
+  month   = apr,
+  review  = {Thorough writing on how to generate an artificial dataset with controllable parameters according to some constraints regarding the covariance matrices of the classes.
+
+Strong results from classifier analysis and nice work on parameter optimisation.},
+  url     = {https://doi.org/10.1371/journal.pone.0094137},
+}
+
+@Comment{jabref-meta: databaseType:bibtex;}
